Fix critical security and correctness issues from cross-model review
- Fix RequestService not passing SourceControlType to BuildDetails,
  which caused all GitHub builds to fail validation (showstopper)
- Fix SSRF: validate GitHub API host against allowlist before sending
  authenticated requests with PAT token
- Fix long-to-int overflow for GitHub run IDs in BuildServerBuildInfo
  and BuildDetail (GitHub IDs exceed int.MaxValue)
- Replace static appsettings.json config with DI-injected IConfiguration
  for testability and container compatibility
- Replace per-call new HttpClient() with IHttpClientFactory to prevent
  socket exhaustion under load
- Add SSRF host validation with configurable GHE allowlist
- Add direct single-run lookup in ValidateBuildAsync when run ID is known
  instead of fetching all runs and filtering in memory
- Add HttpClient timeout (30s) and X-GitHub-Api-Version header
- Log warning when filterPinnedOnly requested for GitHub (not supported)
- Log warning when GitHubToken is not configured
- Fix null guard in DeployableBuildFactory when project not found
- Fix locale-sensitive url.ToLower() with StringComparison.OrdinalIgnoreCase

+++ src/Dorc.Core/BuildServer/IBuildServerClient.cs
@@ -41,7 +41,7 @@ namespace Dorc.Core.BuildServer
         public string BuildUri { get; set; } = string.Empty;
         public string ProjectName { get; set; } = string.Empty;
         public string DefinitionName { get; set; } = string.Empty;
-        public int BuildId { get; set; }
+        public long BuildId { get; set; }
         public string BuildNumber { get; set; } = string.Empty;
         public string ArtifactDownloadUrl { get; set; } = string.Empty;
     }
